refactor: impl preview code view
