+++ lib/src/ui/components/preview_container.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:shirley/src/ui/components/preview/preview_code_view.dart';
 import 'package:shirley/src/ui/components/preview/preview_widget_view.dart';
-import 'package:shirley/src/ui/components/shirley_snack_bar.dart';
 import 'package:syntax_highlight/syntax_highlight.dart';
 
 class PreviewContainer extends HookWidget {
@@ -27,11 +26,7 @@ class PreviewContainer extends HookWidget {
       return;
     }, [theme]);
 
-    final copying = useState(false);
-
     const primaryColor = Color.fromRGBO(228, 23, 73, 1);
-    const secondaryColor = Color.fromRGBO(245, 88, 123, 1);
-    const fourthlyColor = Color.fromRGBO(255, 245, 145, 1);
 
     return DefaultTabController(
       length: 2,
@@ -63,59 +58,7 @@ class PreviewContainer extends HookWidget {
                 physics: const NeverScrollableScrollPhysics(),
                 children: [
                   PreviewWidgetView(json: json),
-                  Stack(
-                    children: [
-                      Padding(
-                        padding: const EdgeInsets.all(16),
-                        child: SelectableText.rich(
-                          hightlighter.value.highlight(code),
-                          style: TextStyle(fontSize: 14),
-                        ),
-                      ),
-                      Positioned(
-                        top: 12,
-                        right: 12,
-                        child: SizedBox(
-                          width: 28,
-                          height: 28,
-                          child: IconButton(
-                            style: IconButton.styleFrom(
-                              foregroundColor: fourthlyColor,
-                              backgroundColor: secondaryColor,
-                            ),
-                            padding: EdgeInsets.all(0.0),
-                            icon: Icon(
-                              copying.value
-                                  ? Icons.hourglass_empty
-                                  : Icons.copy,
-                              size: 20,
-                            ),
-                            onPressed: () async {
-                              copying.value = true;
-
-                              try {
-                                final data = ClipboardData(text: code);
-                                await Clipboard.setData(data);
-
-                                copying.value = false;
-
-                                if (context.mounted) {
-                                  showSnackBar(context, 'Copied Text!!');
-                                }
-                              } catch (e) {
-                                copying.value = false;
-
-                                if (context.mounted) {
-                                  showSnackBar(context, 'Copied Failure',
-                                      isError: true);
-                                }
-                              }
-                            },
-                          ),
-                        ),
-                      )
-                    ],
-                  ),
+                  PreviewCodeView(code: code, theme: theme),
                 ],
               ),
             ),
